Move Scope into its own module.
